Refactor: Modular Architecture for Environment File Handling
